requirements.txt
absl-py==0.7.1
astor==0.7.1
backports.weakref==1.0.post1
enum34==1.1.6
funcsigs==1.0.2
#futures==3.2.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==3.0.4
numpy==1.16.4
protobuf==3.7.1
six==1.10
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
PyDICOM==1.2.2
nose==1.3.7
sphinx==2.1.0
nibabel==2.4.1
matplotlib==3.1.0
pillow==6.0.0
ipykernel==5.1.1 #need to run this: ipython kernel install --user --name=env
glob3==0.0.1
imageio=2.5.0
